Fix clippy warnings and move workspace-level config to root Cargo.toml

+++ faunus/src/analysis/scaled_widom_insertion.rs
@@ -367,7 +367,7 @@ impl<T: Context> Analyze<T> for ScaledWidomInsertion {
                 for j in group.iter_active() {
                     let pos_j = context.position(j);
                     let dist_sq = cell.distance_squared(&r_ghost, &pos_j);
-                    let atom_type = context.atom_kind(j) as usize;
+                    let atom_type = context.atom_kind(j);
 
                     du_sr += self.sr_potentials[atom_type].isotropic_twobody_energy(dist_sq);
